+++ packages/core-components/src/components/LinkButton/LinkButton.test.tsx
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2020 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import { render, fireEvent, act } from '@testing-library/react';
+import { wrapInTestApp } from '@backstage/test-utils';
+import { LinkButton } from './LinkButton';
+import { Route, Routes } from 'react-router-dom';
+
+describe('<LinkButton />', () => {
+  it('navigates using react-router', async () => {
+    const testString = 'This is test string';
+    const linkButtonLabel = 'Navigate!';
+    const { getByText } = render(
+      wrapInTestApp(
+        <>
+          <LinkButton to="/test">{linkButtonLabel}</LinkButton>
+          <Routes>
+            <Route path="/test" element={<p>{testString}</p>} />
+          </Routes>
+        </>,
+      ),
+    );
+
+    expect(() => getByText(testString)).toThrow();
+    await act(async () => {
+      fireEvent.click(getByText(linkButtonLabel));
+    });
+    expect(getByText(testString)).toBeInTheDocument();
+  });
+});