+++ packages/techdocs-cli/src/commands/index.ts
@@ -102,10 +102,6 @@ export function registerCommands(program: Command) {
       'Optional AWS S3 option to force path style.',
     )
     .option(
-      '--awsBucketRootPath',
-      'Optional sub-directory to store files in Amazon S3',
-    )
-    .option(
       '--osCredentialId <OPENSTACK SWIFT APPLICATION CREDENTIAL ID>',
       '(Required for OpenStack) specify when --publisher-type openStackSwift',
     )
@@ -178,6 +174,10 @@ export function registerCommands(program: Command) {
       'Optional AWS S3 option to force path style.',
     )
     .option(
+      '--awsBucketRootPath <AWS BUCKET ROOT PATH>',
+      'Optional sub-directory to store files in Amazon S3',
+    )
+    .option(
       '--osCredentialId <OPENSTACK SWIFT APPLICATION CREDENTIAL ID>',
       '(Required for OpenStack) specify when --publisher-type openStackSwift',
     )
@@ -194,7 +194,7 @@ export function registerCommands(program: Command) {
       '(Required for OpenStack) specify when --publisher-type openStackSwift',
     )
     .option(
-      '--gcsBucketRootPath',
+      '--gcsBucketRootPath <GCS BUCKET ROOT PATH>',
       'Optional sub-directory to store files in Google cloud storage',
     )
     .option(