+++ plugins/scaffolder-backend/src/scaffolder/tasks/StorageTaskBroker.ts
@@ -15,6 +15,7 @@
  */
 
 import { TaskSpec } from '@backstage/plugin-scaffolder-common';
+import { TaskSecrets } from '@backstage/plugin-scaffolder-node';
 import { JsonObject, Observable } from '@backstage/types';
 import { Logger } from 'winston';
 import ObservableImpl from 'zen-observable';
@@ -25,7 +26,6 @@ import {
   TaskBrokerDispatchOptions,
   TaskCompletionState,
   TaskContext,
-  TaskSecrets,
   TaskStore,
 } from './types';