+++ plugins/adr/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @backstage/plugin-adr
 
+## 0.3.1-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.12.4-next.0
+  - @backstage/plugin-search-react@1.5.0-next.0
+  - @backstage/plugin-catalog-react@1.3.0-next.1
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/core-plugin-api@1.3.0
+  - @backstage/integration-react@1.1.10-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-adr-common@0.2.6-next.0
+  - @backstage/plugin-search-common@1.2.1
+
+## 0.3.1-next.0
+
+### Patch Changes
+
+- 496cf6aff0: Render the common `<MissingAnnotationEmptyState />` component when the `backstage.io/adr-location` annotation is missing from the component
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.3.0-next.0
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/plugin-adr-common@0.2.6-next.0
+  - @backstage/integration-react@1.1.9
+
 ## 0.3.0
 
 ### Minor Changes