+++ plugins/catalog-backend-module-aws/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @backstage/plugin-catalog-backend-module-aws
 
+## 0.1.15-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-backend@1.7.2-next.1
+  - @backstage/backend-common@0.18.2-next.1
+  - @backstage/backend-plugin-api@0.3.2-next.1
+  - @backstage/backend-tasks@0.4.3-next.1
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/config@1.0.6
+  - @backstage/errors@1.1.4
+  - @backstage/integration@1.4.2
+  - @backstage/types@1.0.2
+  - @backstage/plugin-catalog-node@1.3.3-next.1
+
+## 0.1.15-next.0
+
+### Patch Changes
+
+- 8e025f1347: Added support for `externalId` when assuming role in `AwsS3EntityProvider`
+- Updated dependencies
+  - @backstage/catalog-model@1.1.6-next.0
+  - @backstage/backend-common@0.18.2-next.0
+  - @backstage/plugin-catalog-backend@1.7.2-next.0
+  - @backstage/plugin-catalog-node@1.3.3-next.0
+  - @backstage/backend-tasks@0.4.3-next.0
+  - @backstage/backend-plugin-api@0.3.2-next.0
+
 ## 0.1.13
 
 ### Patch Changes